+++ R/form_assignment_revoke.R
@@ -1,0 +1,76 @@
+#' Revoke a Form Role Assignment from an Actor.
+#'
+#' `r lifecycle::badge("experimental")`
+#'
+#' Given a Role and an Actor, this endpoint unassigns that Role from that
+#' Actor for this particular Form.
+#'
+#' @template param-pid
+#' @template param-fid
+#' @param role_id (character or numeric) The Role ID, typically the integer
+#'   ID of the Role.
+#'   A Role system name can also be supplied if the Role has one.
+#' @param actor_id (numeric) The integer ID of the Actor.
+#' @template param-url
+#' @template param-auth
+#' @template param-retries
+#' @return A list with a single element `success` (`TRUE`) as per the ODK
+#'   Central API docs.
+# nolint start
+#' @seealso \url{https://docs.getodk.org/central-api-form-management/#revoking-a-form-role-assignment-from-an-actor}
+# nolint end
+#' @family form-management
+#' @export
+#' @examples
+#' \dontrun{
+#' # See vignette("setup") for setup and authentication options
+#' # ruODK::ru_setup(svc = "....svc", un = "me@email.com", pw = "...")
+#'
+#' fl <- form_list()
+#'
+#' form_assignment_revoke(
+#'   fid = fl$fid[[1]],
+#'   role_id = "manager",
+#'   actor_id = 14
+#' )
+#' # > $success
+#' # > [1] TRUE
+#' }
+form_assignment_revoke <- function(
+  pid = get_default_pid(),
+  fid = get_default_fid(),
+  role_id,
+  actor_id,
+  url = get_default_url(),
+  un = get_default_un(),
+  pw = get_default_pw(),
+  retries = get_retries()
+) {
+  yell_if_missing(url, un, pw, pid = pid, fid = fid)
+
+  if (missing(role_id) || length(role_id) != 1L || is.na(role_id)) {
+    ru_msg_abort("role_id must be a single Role ID.")
+  }
+  if (missing(actor_id) || length(actor_id) != 1L || is.na(actor_id)) {
+    ru_msg_abort("actor_id must be a single Actor ID.")
+  }
+
+  httr::RETRY(
+    "DELETE",
+    httr::modify_url(
+      url,
+      path = glue::glue(
+        "v1/projects/{pid}/forms/",
+        "{URLencode(fid, reserved = TRUE)}/assignments/{role_id}/{actor_id}"
+      )
+    ),
+    httr::add_headers("Accept" = "application/json"),
+    httr::authenticate(un, pw),
+    times = retries
+  ) |>
+    yell_if_error(url, un, pw) |>
+    httr::content(encoding = "utf-8") |>
+    janitor::clean_names()
+}
+
+# usethis::use_test("form_assignment_revoke")  # nolint